corrected data type

+++ alx_book_store.sql
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
     order_details INT PRIMARY KEY,
     order_id INT,
     book_id INT,
-    quantity INT,
+    quantity DOUBLE,
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
